Major updates to the url structure and "add task"
Changed url structure to have tasks (and add task instances) to be
stored in the project path that they are associated with.
Updated the add task page to allow creation of new tasks, editing
existing ones, and deleting existing ones. Actions taken from the add
task page now return to the project page they linked from rather than
the user’s dashboard.

+++ TaskMaster/taskmaster_app/views.py
@@ -4,6 +4,7 @@ from django.core.urlresolvers import reverse
 from django.http import HttpResponseRedirect
 from django.utils.text import slugify
 from django.contrib.auth.decorators import login_required
+from django.utils import timezone
 
 from taskmaster_app.models import Project, Task, Resource
 from taskmaster_app.forms import ProjectForm, TaskForm
@@ -44,34 +45,42 @@ def add_project(request):
     return render(request, 'taskmaster_app/addproject.html',{'form':form, 'project':project})
 
 @login_required(login_url='/')
-def add_task(request):
+def add_task(request, **kwargs):
+    slug = kwargs['project']
+    project = get_object_or_404(Project, slug=slug)
+    
     id = request.GET.get('id', None)
     if id is not None:
         task = get_object_or_404(Task, id=id)
     else:
-        task = None
+        task = Task(title='',task_instructions='',project=project,deadline=timezone.now())
         
     if request.method == 'POST':
+        if request.POST.get('control') == 'cancel':
+            return HttpResponseRedirect(reverse('taskmaster_app:project_view', kwargs={'project':project.slug}))
         if request.POST.get('control') == 'delete':
             task.delete()
             messages.add_message(request, messages.INFO, 'Task deleted!')
-            return HttpResponseRedirect(reverse('taskmaster_app:dashboard'))
+            return HttpResponseRedirect(reverse('taskmaster_app:project_view', kwargs={'project':project.slug}))
             
-        form = ProjectForm(request.POST)
+        form = TaskForm(request.POST)
         if form.is_valid():
-            t = form.save(commit=False)
-            t.slug = slugify(t.project.title + '_' + t.title)
-            t.save()
-            messages.add_message(request, messages.INFO, 'Task added!')
+            t = form.save()
+            t.slug = slugify(t.title)
+            t.save()            
+            if id is None:
+                messages.add_message(request, messages.INFO, 'Task added!')
+            else:
+                messages.add_message(request, messages.INFO, 'Task saved!')
             return HttpResponseRedirect(reverse('taskmaster_app:dashboard'))
     else:
         form = TaskForm(instance=task)
     
-    return render(request, 'taskmaster_app/addtask.html',{'form':form, 'task':task})
+    return render(request, 'taskmaster_app/addtask.html',{'form':form, 'task':task, 'Tasks':Task.objects.all()})
 
 @login_required(login_url='/')
 def project_view(request, **kwargs):
-    slug = kwargs['slug']
+    slug = kwargs['project']
     project = get_object_or_404(Project, slug=slug)
     tasks = project.tasks.all()
     resources = project.resources.all()
@@ -80,7 +89,7 @@ def project_view(request, **kwargs):
 
 @login_required(login_url='/')
 def task_view(request, **kwargs):
-    slug = kwargs['slug']
+    slug = kwargs['task']
     task = get_object_or_404(Task, slug=slug)
     project = task.project
     resources = task.resources.all()
@@ -89,7 +98,7 @@ def task_view(request, **kwargs):
     
 @login_required(login_url='/')
 def resource_view(request, **kwargs):
-    slug = kwargs['slug']
+    slug = kwargs['resource']
     resource = get_object_or_404(Resource, slug=slug)
     projects = resource.projects.all()
     tasks = resource.tasks.all()
